requirements.txt
streamlit==1.38.0
pytesseract==0.3.13
opencv-python==4.10.0.84
Pillow==10.4.0
numpy==2.1.1
rapidfuzz==3.9.7
PyMuPDF==1.24.9
pdf2image==1.17.0
reportlab==4.2.2
pydantic==2.9.2
